refactor(接口测试): 重构APIBackendListenerClient，结果处理和存储分离
