+++ backend/src/main/java/io/metersphere/api/service/ApiAutomationService.java
@@ -134,8 +134,12 @@ public class ApiAutomationService {
         return apiTestEnvironmentMapper.selectByPrimaryKey(id);
     }
 
-    public User getUser(String id) {
-        return userMapper.selectByPrimaryKey(id);
+    public String getUser(String id) {
+        User user = userMapper.selectByPrimaryKey(id);
+        if (user != null) {
+            return user.getName();
+        }
+        return null;
     }
 
     public List<ApiScenarioDTO> list(ApiScenarioRequest request) {
@@ -1572,7 +1576,7 @@ public class ApiAutomationService {
         ApiScenarioWithBLOBs apiScenarioWithBLOBs = new ApiScenarioWithBLOBs();
         BeanUtils.copyBean(apiScenarioWithBLOBs, request);
         apiScenarioWithBLOBs.setUpdateTime(System.currentTimeMillis());
-        if(apiScenarioWithBLOBs.getScenarioDefinition() != null){
+        if (apiScenarioWithBLOBs.getScenarioDefinition() != null) {
             List<ApiMethodUrlDTO> useUrl = this.parseUrl(apiScenarioWithBLOBs);
             apiScenarioWithBLOBs.setUseUrl(JSONArray.toJSONString(useUrl));
         }
@@ -1808,11 +1812,11 @@ public class ApiAutomationService {
      * 1.场景中复制的接口
      * 2.场景中引用/复制的案例
      * 3.场景中的自定义路径与接口定义中的匹配
-     *
+     * <p>
      * 匹配场景中用到的路径
      *
-     * @param allScenarioInfoList     场景集合（id / scenario大字段 必须有数据）
-     * @param allEffectiveApiList     接口集合（id / path 必须有数据）
+     * @param allScenarioInfoList 场景集合（id / scenario大字段 必须有数据）
+     * @param allEffectiveApiList 接口集合（id / path 必须有数据）
      * @return
      */
     public float countInterfaceCoverage(List<ApiScenarioWithBLOBs> allScenarioInfoList, List<ApiDefinition> allEffectiveApiList) {
@@ -1821,20 +1825,20 @@ public class ApiAutomationService {
             return 100;
         }
 
-        Map<ApiMethodUrlDTO,List<String>> urlMap = new HashMap<>();
+        Map<ApiMethodUrlDTO, List<String>> urlMap = new HashMap<>();
         for (ApiDefinition model : allEffectiveApiList) {
             String url = model.getPath();
             String method = model.getMethod();
             String id = model.getId();
 
-            ApiMethodUrlDTO dto = new ApiMethodUrlDTO(url,method);
+            ApiMethodUrlDTO dto = new ApiMethodUrlDTO(url, method);
 
-            if(urlMap.containsKey(dto)){
+            if (urlMap.containsKey(dto)) {
                 urlMap.get(dto).add(id);
-            }else{
+            } else {
                 List<String> list = new ArrayList<>();
                 list.add(id);
-                urlMap.put(dto,list);
+                urlMap.put(dto, list);
             }
         }
 
@@ -1845,9 +1849,9 @@ public class ApiAutomationService {
         List<ApiMethodUrlDTO> urlList = new ArrayList<>();
         for (ApiScenarioWithBLOBs model : allScenarioInfoList) {
             List<ApiMethodUrlDTO> useUrl = this.getScenarioUseUrl(model);
-            if(CollectionUtils.isNotEmpty(useUrl)){
-                for (ApiMethodUrlDTO dto :useUrl) {
-                    if(!urlList.contains(dto)){
+            if (CollectionUtils.isNotEmpty(useUrl)) {
+                for (ApiMethodUrlDTO dto : useUrl) {
+                    if (!urlList.contains(dto)) {
                         urlList.add(dto);
                     }
                 }
@@ -1868,8 +1872,8 @@ public class ApiAutomationService {
         }
 
         int allApiIdCount = 0;
-        for (List<String> allApiIdList:urlMap.values()){
-            if(CollectionUtils.isNotEmpty(allApiIdList)){
+        for (List<String> allApiIdList : urlMap.values()) {
+            if (CollectionUtils.isNotEmpty(allApiIdList)) {
                 allApiIdCount += allApiIdList.size();
             }
         }
@@ -1881,8 +1885,8 @@ public class ApiAutomationService {
     private List<ApiMethodUrlDTO> getScenarioUseUrl(ApiScenarioWithBLOBs model) {
         List<ApiMethodUrlDTO> useUrlList = new ArrayList<>();
         try {
-            useUrlList = JSONArray.parseArray(model.getUseUrl(),ApiMethodUrlDTO.class);
-        }catch (Exception e){
+            useUrlList = JSONArray.parseArray(model.getUseUrl(), ApiMethodUrlDTO.class);
+        } catch (Exception e) {
         }
         return useUrlList;
     }
@@ -1893,9 +1897,9 @@ public class ApiAutomationService {
         try {
             String scenarioDefiniton = scenario.getScenarioDefinition();
             JSONObject scenarioObj = JSONObject.parseObject(scenarioDefiniton);
-            List<ApiMethodUrlDTO> stepUrlList  = this.getMethodUrlDTOByHashTreeJsonObj(scenarioObj);
-            if(CollectionUtils.isNotEmpty(stepUrlList)){
-                Collection unionList = CollectionUtils.union(urlList,stepUrlList);
+            List<ApiMethodUrlDTO> stepUrlList = this.getMethodUrlDTOByHashTreeJsonObj(scenarioObj);
+            if (CollectionUtils.isNotEmpty(stepUrlList)) {
+                Collection unionList = CollectionUtils.union(urlList, stepUrlList);
                 urlList = new ArrayList<>(unionList);
             }
         } catch (Exception e) {
@@ -1910,22 +1914,22 @@ public class ApiAutomationService {
             JSONArray hashArr = obj.getJSONArray("hashTree");
             for (int i = 0; i < hashArr.size(); i++) {
                 JSONObject elementObj = hashArr.getJSONObject(i);
-                if(elementObj == null){
+                if (elementObj == null) {
                     continue;
                 }
                 if (elementObj.containsKey("url") && elementObj.containsKey("method")) {
                     String url = elementObj.getString("url");
                     String method = elementObj.getString("method");
-                    ApiMethodUrlDTO dto = new ApiMethodUrlDTO(url,method);
-                    if(!returnList.contains(dto)){
+                    ApiMethodUrlDTO dto = new ApiMethodUrlDTO(url, method);
+                    if (!returnList.contains(dto)) {
                         returnList.add(dto);
                     }
                 }
                 if (elementObj.containsKey("path") && elementObj.containsKey("method")) {
                     String path = elementObj.getString("path");
                     String method = elementObj.getString("method");
-                    ApiMethodUrlDTO dto = new ApiMethodUrlDTO(path,method);
-                    if(!returnList.contains(dto)){
+                    ApiMethodUrlDTO dto = new ApiMethodUrlDTO(path, method);
+                    if (!returnList.contains(dto)) {
                         returnList.add(dto);
                     }
                 }
@@ -1933,28 +1937,28 @@ public class ApiAutomationService {
                 if (elementObj.containsKey("id") && elementObj.containsKey("refType")) {
                     String refType = elementObj.getString("refType");
                     String id = elementObj.getString("id");
-                    if(StringUtils.equals("CASE",refType)){
+                    if (StringUtils.equals("CASE", refType)) {
                         ApiDefinition apiDefinition = apiTestCaseService.findApiUrlAndMethodById(id);
-                        if(apiDefinition != null){
-                            ApiMethodUrlDTO dto = new ApiMethodUrlDTO(apiDefinition.getPath(),apiDefinition.getMethod());
-                            if(!returnList.contains(dto)){
+                        if (apiDefinition != null) {
+                            ApiMethodUrlDTO dto = new ApiMethodUrlDTO(apiDefinition.getPath(), apiDefinition.getMethod());
+                            if (!returnList.contains(dto)) {
                                 returnList.add(dto);
                             }
                         }
-                    }else if(StringUtils.equals("API",refType)){
+                    } else if (StringUtils.equals("API", refType)) {
                         ApiDefinition apiDefinition = apiDefinitionService.selectUrlAndMethodById(id);
-                        if(apiDefinition != null){
-                            ApiMethodUrlDTO dto = new ApiMethodUrlDTO(apiDefinition.getPath(),apiDefinition.getMethod());
-                            if(!returnList.contains(dto)){
+                        if (apiDefinition != null) {
+                            ApiMethodUrlDTO dto = new ApiMethodUrlDTO(apiDefinition.getPath(), apiDefinition.getMethod());
+                            if (!returnList.contains(dto)) {
                                 returnList.add(dto);
                             }
                         }
                     }
                 }
 
-                List<ApiMethodUrlDTO> stepUrlList  = this.getMethodUrlDTOByHashTreeJsonObj(elementObj);
-                if(CollectionUtils.isNotEmpty(stepUrlList)){
-                    Collection unionList = CollectionUtils.union(returnList,stepUrlList);
+                List<ApiMethodUrlDTO> stepUrlList = this.getMethodUrlDTOByHashTreeJsonObj(elementObj);
+                if (CollectionUtils.isNotEmpty(stepUrlList)) {
+                    Collection unionList = CollectionUtils.union(returnList, stepUrlList);
                     returnList = new ArrayList<>(unionList);
                 }
             }
@@ -1987,7 +1991,6 @@ public class ApiAutomationService {
         } catch (Exception e) {
         }
     }
-
 
 
     public ScenarioEnv getApiScenarioProjectId(String id) {
@@ -2092,11 +2095,11 @@ public class ApiAutomationService {
     public void checkApiScenarioUseUrl() {
         List<String> noUrlScenarioIdList = extApiScenarioMapper.selectIdsByUseUrlIsNull();
         for (String id : noUrlScenarioIdList) {
-            try{
+            try {
                 ApiScenarioWithBLOBs scenario = apiScenarioMapper.selectByPrimaryKey(id);
-                if(scenario.getUseUrl() == null){
+                if (scenario.getUseUrl() == null) {
                     List<ApiMethodUrlDTO> useUrl = this.parseUrl(scenario);
-                    if(useUrl != null){
+                    if (useUrl != null) {
                         ApiScenarioWithBLOBs updateModel = new ApiScenarioWithBLOBs();
                         updateModel.setId(scenario.getId());
                         updateModel.setUseUrl(JSONArray.toJSONString(useUrl));
@@ -2105,7 +2108,7 @@ public class ApiAutomationService {
                     }
                 }
                 scenario = null;
-            }catch (Exception e){
+            } catch (Exception e) {
                 e.printStackTrace();
             }
